Added bootstrap and clean scripts
git-svn-id: svn+ssh://svn.gersteinlab.org/repos/bioinfo/fusionSeq@618 2fd4de08-f6cc-4d2b-b124-baeea9ba92f3

+++ src/sqvUtil.c
@@ -26,7 +26,7 @@ int inRegion (int istart, int iend, int rstart, int rend)
   return 0;
 }
 
-unsigned int randomN (int n)
+unsigned long int randomN (int n)
 {
   return random() / ((2 << 30) - 1) * n;
 }
