+++ src/sqvUtil.h
@@ -55,7 +55,7 @@ typedef struct {
 
 
 void swapInt (int *a, int *b);
-unsigned int randomN (int n);
+unsigned long int randomN (int n);
 
 int inRegion (int istart, int iend, int rstart, int rend);
 int inRegions (Array regions, int chromosome, int start, int end);
